+++ NEWS.md
@@ -1,11 +1,77 @@
 
-# IEATools 0.1.53 (2021-05-07)
+# IEATools 0.1.58 (2021-08-20)
+
+* Changed "Epsilon"" matrix name to "Balancing".
+  The old name was causing issues in ECCTools.
+
+# IEATools 0.1.57 (2021-08-19) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5222003.svg)](https://doi.org/10.5281/zenodo.5222003)
+
+* Fixed a bug where missing EIOU was not properly detected.
+  It could be missing or `NULL`, but `NULL` was not detected.
+* No new tests.
+    * Still at 833 tests, all passing.
+    * Test coverage remains at 100 %.
+
+
+# IEATools 0.1.56 (2021-08-18) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5217278.svg)](https://doi.org/10.5281/zenodo.5217278)
+
+* `extend_to_useful()` now robust to cases where there is no EIOU,
+  thanks to now using `matsindf::matsindf_apply()`.
+* Now using `matsindf::matsindf_apply()` for `extend_to_useful()`.
+  This change will enable better handling of cases where
+  a country has no EIOU.
+  The first use case will be bunkers.
+* No longer cleaning matrices in `extend_to_useful_helper()`, 
+  because some vectors may be the `0` vector and be eliminated.
+* Added `replace_null_UR()` function to the workflow in the `prep_psut()` function.
+* New function `replace_null_UR()` replaces missing or `NULL` 
+  `R`, `U_feed`, `U_EIOU`, `U`, and `r_EIOU`
+  with **0** matrices with appropriate row and column names.
+  The replacements are built from **Y** and **V** matrices.
+  The need for this functionality arises when
+  the last stage is final energy 
+  and imports (**V** matrix) 
+  are the only source of an energy carrier that 
+  is consumed in final demand (**Y** matrix).
+  In that situation, the **R** and **U** matrices 
+  will be missing, 
+  and they can be replaced by **0** matrices with the right dimensions.
+* In the specification process,
+  the `Flow` for countries `World_X_bunkers` are now
+  specified to be "International navigation" and "International aviation"
+  to correspond with their domestic equivalents.
+* New tests for new code.
+    * Up to 833 tests, all passing.
+    * Test coverage remains at 100 %.
+
+
+# IEATools 0.1.55 (2021-07-21) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5119070.svg)](https://doi.org/10.5281/zenodo.5119070)
+
+* Modifications to code and tests
+  to accommodate World marine bunkers and World aviation bunkers being their own country.
+* New tests for new features.
+    * Up to 808 tests, all passing.
+    * Test coverage remains at 100 %.
+
+
+# IEATools 0.1.54 (2021-07-09) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.5086372.svg)](https://doi.org/10.5281/zenodo.5086372)
+
+* Modifications to enable per-machine
+  final-to-useful efficiency data.
+* New tests for new features.
+    * Up to 790 tests, all passing.
+    * Test coverage remains at 100 %.
+
+
+# IEATools 0.1.53 (2021-06-08)
+>>>>>>> hotfix-0.1.58
 
 * The "Oil and gas extraction" activity is split in "Oil extraction",
   which extracts oil products, and "Natural gas extraction", which extracts natural gas.
 * New `split_oil_gas_extraction_eiou()` function splits the EIOU of 
   the "Oil and gas extraction" industry into EIOU for the
   "Oil extraction" and for the "Natural gas extraction" activities.
+
 
 # IEATools 0.1.52 (2021-05-07)